Move BrowserWebsocket to util/
