+++ lib/gateway/Shard.js
@@ -12,7 +12,7 @@ const ExtendedUser = require("../structures/ExtendedUser");
 const User = require("../structures/User");
 const Invite = require("../structures/Invite");
 const Constants = require("../Constants");
-let WebSocket = typeof window !== "undefined" ? require("./BrowserWebsocket") : require("ws");
+let WebSocket = typeof window !== "undefined" ? require("../util/BrowserWebsocket") : require("ws");
 
 let EventEmitter;
 try {
