+++ lib/util/BrowserWebsocket.js
