+++ lib/voice/VoiceConnection.js
@@ -8,7 +8,7 @@ const Piper = require("./Piper");
 const VoiceDataStream = require("./VoiceDataStream");
 const {createOpus} = require("../util/Opus");
 
-let WebSocket = typeof window !== "undefined" ? require("../gateway/BrowserWebsocket") : require("ws");
+let WebSocket = typeof window !== "undefined" ? require("../util/BrowserWebsocket") : require("ws");
 
 let EventEmitter;
 try {
